server/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
pdfplumber
pymupdf
python-docx
google-generativeai
weasyprint
xhtml2pdf
